Moves common config to conf.py, and app specific config to config_app.py
